Move `KeyPair` back to `stark_curve_signer.py`
